@@ -78,7 +78,7 @@ class ParentParsers:
 
     list_parser = argparse.ArgumentParser(add_help=False)
     list_parser.add_argument(
-        "--package_name",
+        "--package-name",
         "-p",
         type=str,
         default="all",
@@ -103,6 +103,30 @@ class ParentParsers:
                 For more in formation on headers available, run
                 'geoips list <cmd> <positional_args> --columns help'.
                 """,
+    )
+
+    config_parser = argparse.ArgumentParser(add_help=False)
+    config_parser.add_argument(
+        "-n",
+        "--namespace",
+        default="geoips.plugin_packages",
+        type=str,
+        help=(
+            "The namespace of plugin packages to create plugin registries for. "
+            "If not specified, this defaults to 'geoips.plugin_packages'."
+        ),
+    )
+    config_parser.add_argument(
+        "-p",
+        "--packages",
+        default=None,
+        nargs="+",
+        type=str,
+        help=(
+            "The plugin packages to create or delete plugin registries for. Defaults to"
+            " None. If None, all plugin packages under 'namespace' will have their "
+            "plugin registries created or deleted, based on the command supplied."
+        ),
     )
 
 
@@ -162,6 +186,18 @@ class GeoipsCommand(abc.ABC):
             curr_parent = self.parent
             self.parent_parsers = []
             while curr_parent and hasattr(ParentParsers, f"{curr_parent.name}_parser"):
+                # Don't add parent parser arguments to 'geoips config install'. Config's
+                # shared arguments only apply to 'create-registries' and
+                # 'delete-registries' and the time being. I've looked for easier
+                # workarounds and can't find any suitable options other than hardcoding.
+
+                # If we end up needing to add more hardcoded special cases here, it
+                # might be worth looking into other methods of how to share arguments
+                # among specific commands, or subsets of commands. For now, this works
+                if self.name == "install" and self.parent.name == "config":
+                    break
+                if self.name == "install-github" and self.parent.name == "config":
+                    break
                 self.parent_parsers.insert(
                     0,
                     getattr(ParentParsers, f"{curr_parent.name}_parser"),
