@@ -1,3 +1,6 @@
+# # # This source code is subject to the license referenced at
+# # # https://github.com/NRLMMD-GEOIPS.
+
 """Pydantic base models for GeoIPS.
 
 Intended for use by other base models.
@@ -10,7 +13,8 @@ Other models defined here validate field types within child plugin models.
 # Python Standard Libraries
 import keyword
 import logging
-from typing import Union, Tuple
+from typing import ClassVar, Union, Tuple
+import warnings
 
 # Third-Party Libraries
 from pydantic import (
@@ -23,10 +27,13 @@ from pydantic import (
 from pydantic_core import PydanticCustomError
 from pydantic.functional_validators import AfterValidator
 from typing_extensions import Annotated
-
+from pydantic._internal._model_construction import (
+    ModelMetaclass,
+)  # internal API, but safe to use
 
 # GeoIPS imports
 from geoips import interfaces
+from geoips.geoips_utils import get_interface_module
 
 LOG = logging.getLogger(__name__)
 
@@ -37,8 +44,8 @@ ColorType = Union[ColorTuple, str]
 class PrettyBaseModel(BaseModel):
     """Make Pydantic models pretty-print by default.
 
-    This model overrides the default string representation of Pyantic models to generate
-    a user-friendly, JSON-formatted output with two-space indentation.
+    This model overrides the default string representation of Pydantic models to
+    generate a user-friendly, JSON-formatted output with two-space indentation.
     """
 
     def __str__(self) -> str:
@@ -142,25 +149,56 @@ def python_identifier(val: str) -> str:
 PythonIdentifier = Annotated[str, AfterValidator(python_identifier)]
 
 
-def get_interfaces() -> set[str]:
+def get_interfaces(namespace) -> set[str]:
     """Return a set of distinct interfaces.
 
     This function returns all available plugin interfaces. The results are cached for
-    runtime memory optimizaiton.
+    runtime memory optimization.
 
     Returns
     -------
     set of str
         set of interfaces
     """
-    return {
-        available_interfaces
-        for ifs in interfaces.list_available_interfaces().values()
-        for available_interfaces in ifs
-    }
+    if namespace == "geoips.plugin_packages":
+        return {
+            available_interfaces
+            for ifs in interfaces.list_available_interfaces().values()
+            for available_interfaces in ifs
+        }
+    else:
+        mod = get_interface_module(namespace)
+        return set(mod.__all__)
 
 
-class PluginModel(FrozenModel):
+class PluginModelMetadata(ModelMetaclass):
+    """API version and namespace metadata for the corresponding plugin model.
+
+    This is used to derive 'apiVersion' and 'namespace' for any given PluginModel.
+    PluginModel can be instantiated directly or a child class of PluginModel can be
+    instantiated and the functionality will for the same.
+
+    Initially attempted to use __init_subclass__ in the PluginModel class itself, but
+    that only supported child classes of PluginModel (i.e. WorkflowPluginModel, ...),
+    but not instantiation of PluginModel itself.
+
+    NOTE: Need to inherit from ModelMetaclass, otherwise we'll wind up with this error:
+
+    E TypeError: metaclass conflict: the metaclass of a derived class must be a
+    (non-strict) subclass of the metaclasses of all its bases
+    """
+
+    def __new__(mcs, name, bases, namespace, **kwargs):
+        """Instantiate a new PluginModelMetadata class."""
+        cls = super().__new__(mcs, name, bases, namespace)
+        # Set apiVersion if not already set
+        if not hasattr(cls, "apiVersion") or cls.apiVersion is None:
+            cls.apiVersion = "geoips/v1"
+        cls._namespace = f"{cls.apiVersion.split('/')[0]}.plugin_packages"
+        return cls
+
+
+class PluginModel(FrozenModel, metaclass=PluginModelMetadata):
     """Base Plugin model for all GeoIPS plugins.
 
     This should be used as the base class for all top-level
@@ -171,6 +209,9 @@ class PluginModel(FrozenModel):
     <https://github.com/NRLMMD-GEOIPS/geoips/blob/main/docs/source/tutorials/plugin_development/product_default.rst>`_
     for more information about how this is used.
     """
+
+    apiVerson: ClassVar[str | None] = None
+    _namespace: ClassVar[str | None] = None
 
     interface: PythonIdentifier = Field(
         ...,
@@ -213,17 +254,21 @@ class PluginModel(FrozenModel):
         # name is guaranteed to exist due to Pydantic validation.
         # No need to raise an error for 'name'.
         interface_name = values.get("interface")
+        if cls._namespace == "geoips.plugin_packages":
+            ints = interfaces
+        else:
+            ints = get_interface_module(cls._namespace)
         try:
-            metadata = getattr(interfaces, interface_name).get_plugin_metadata(
+            metadata = getattr(ints, interface_name).get_plugin_metadata(
                 values.get("name")
             )
         except AttributeError:
             raise ValueError(
                 f"Invalid interface: '{interface_name}'."
-                f"Must be one of {get_interfaces()}"
+                f"Must be one of {get_interfaces(cls._namespace)}"
             )
         # the above exception handling would be further improved by checking the
-        # existence of plugin registry in the fuutre issue #906
+        # existence of plugin registry in the future issue #906
         if "package" not in metadata:
             err_msg = (
                 "Metadata for '%s' workflow plugin must contain 'package' key."
@@ -244,8 +289,6 @@ class PluginModel(FrozenModel):
 
         Parameters
         ----------
-        cls : Type
-            PluginModel class.
         value :
             Value of the 'interface' field to validate.
 
@@ -259,9 +302,9 @@ class PluginModel(FrozenModel):
         ValueError
             If the 'interface' field value is not in the list of valid interfaces.
         """
-        valid_interfaces = get_interfaces()
+        valid_interfaces = get_interfaces(cls._namespace)
         if value not in valid_interfaces:
-            err_msg = f"Incorrect interface:'{value}'.Must be one of {valid_interfaces}"
+            err_msg = f"Invalid interface:'{value}'. Must be one of {valid_interfaces}"
             LOG.critical(err_msg, exc_info=True)
             raise ValueError(err_msg)
         return value
@@ -271,7 +314,7 @@ class PluginModel(FrozenModel):
         cls: type["PluginModel"], values: dict[str, str | int | float | None]
     ):
         """
-        Set ``description`` to first line of ``dosctring`` field if not provided.
+        Set ``description`` to first line of ``docstring`` field if not provided.
 
         Parameters
         ----------
@@ -312,10 +355,12 @@ class PluginModel(FrozenModel):
         str
             Validated ``description`` string.
 
-        Raises
-        ------
-        PydanticCustomError
+        Warns
+        -----
+        FutureWarning
             If the ``description`` field violates any of the validation rules.
+            This will raise a ValidationError in a future release.
+
         """
         error_messages = {
             "single_line": "Description must be a single line.",
@@ -324,27 +369,37 @@ class PluginModel(FrozenModel):
             ),
             "length_error": "Description cannot be more than 72 characters, reduce by:",
         }
-        if "\n" in value:
-            LOG.critical(
-                "'error': %s 'input_provided': %r",
-                error_messages["single_line"],
-                value,
-                exc_info=True,
+        try:
+
+            def _log_and_raise_custom_pydantic_errors(
+                error_type: str, error_message: str, value: str
+            ):
+                warnings.warn(
+                    f"'error': {error_message} 'input_provided': {value}",
+                    FutureWarning,
+                    stacklevel=3,
+                )
+                raise PydanticCustomError(error_type, error_message)
+
+            if "\n" in value:
+                _log_and_raise_custom_pydantic_errors(
+                    "single_line", error_messages["single_line"], value
+                )
+            elif not value or not value[0].isalnum() or not value.endswith("."):
+                _log_and_raise_custom_pydantic_errors(
+                    "format_error", error_messages["format_error"], value
+                )
+            elif len(value) > 72:
+                excess_length = len(value) - 72
+                err_msg = f"{error_messages['length_error']} {excess_length} characters"
+                _log_and_raise_custom_pydantic_errors("length_error", err_msg, value)
+
+        except PydanticCustomError as e:
+            warnings.warn(
+                f"Future ValidationError encountered. This current warning will become "
+                f"an error in a future release. {e}",
+                FutureWarning,
+                stacklevel=2,
             )
-            raise PydanticCustomError("single_line", error_messages["single_line"])
-        if not (value[0].isalnum() and value.endswith(".")):
-            LOG.critical(
-                "'error': %s 'input_provided': %r",
-                error_messages["format_error"],
-                value,
-                exc_info=True,
-            )
-            raise PydanticCustomError("format_error", error_messages["format_error"])
-        if len(value) > 72:
-            excess_length = len(value) - 72
-            err_msg = f"{error_messages['length_error']} {excess_length} characters"
-            LOG.critical(
-                "'error': %s 'input_provided': %r", err_msg, value, exc_info=True
-            )
-            raise PydanticCustomError("length_error", err_msg)
+
         return value
